similarity_testing/requirements.txt
Flask==2.2.2
sentence-transformers==2.2.2
scikit-learn==1.2.2
transformers==4.27.0
newspaper3k==0.2.8